Allow persisting multiple content pieces
When compacting manifest data, allow persisting multiple content pieces.
This should be alright because flushes on the index are already
disabled, which means all changes should appear atomically.

Uses a max number of manifests per content piece as that's easily
available and can be used as a proxy for total content size to some
extent.