Further harden multi-content write path
Default to having the feature off if the env var isn't set or is set to
a value that is <= 0. This makes things a bit safer since it requires
explicit configuration to enable.

Also be more careful about the types being used to contain values since
we're parsing data from a string which may make different assumptions
about integer widths. This also adds a layer of protection in some sense
since values larger than maxInt64 will overflow to a negative value and
result in the feature being disabled instead of writing out 1 manifest
per content piece.